@@ -1,4 +1,4 @@
-module.exports = {
+export default {
     displayName: 'Window on the right',
 
     hook: 'menu',
@@ -6,14 +6,14 @@ module.exports = {
     /*
         List of classnames that will be attached to your
         plugin upon mounting
-	*/
-    className: 'plugin-rhpane plugin-mobile-rhpane',
+    */
+    className: 'plugin-rhpane',
 
     /*
         Identifier of a pane, where your plugin will reside.
         Since we want to have only one pane on the right side,
         this string will guarantee, that all other plugins
         on the right side will be closed.
-	*/
-    exclusive: 'rhpane',
+    */
+    pane: 'rhpane',
 };
