@@ -1,8 +1,8 @@
-module.exports = {
+export default {
     displayName: 'Multiple formats',
     hook: 'menu',
     className: 'plugin-rhpane plugin-mobile-rhpane',
-    exclusive: 'rhpane',
+    pane: 'rhpane',
     dependencies: [
         'https://cdnjs.cloudflare.com/ajax/libs/leaflet-omnivore/0.3.4/leaflet-omnivore.min.js',
     ],
