@@ -1,4 +1,4 @@
-module.exports = {
+export default {
     displayName: 'Interpolator',
     hook: 'menu',
 };
