@@ -1,6 +1,6 @@
-module.exports = {
+export default {
     displayName: 'Loading backend data',
     hook: 'menu',
     className: 'plugin-lhpane plugin-mobile-fullscreen',
-    exclusive: 'lhpane',
+    pane: 'lhpane',
 };
