@@ -1,7 +1,7 @@
-module.exports = {
+export default {
     displayName: 'drag demo',
     hook: 'contextmenu',
     className: 'plugin-lhpane',
     classNameMobile: 'plugin-mobile-bottom-slide',
-    exclusive: 'lhpane',
+    pane: 'lhpane',
 };
