@@ -1,7 +1,7 @@
-import * as _ from '@windy/utils';
+import { clone } from '@windy/utils';
 
 // Remove data points with some null values
-function validateData(data) {
+export function validateData(data) {
     for (let i = 0; i < data.length; ) {
         const keys = Object.keys(data[i]);
         if (keys.find(key => data[i][key] == null)) {
@@ -13,21 +13,21 @@ function validateData(data) {
 }
 
 // Use when float inaccuracy may occur
-const almostEqual = (a, b) => Math.abs(a - b) < 0.000001;
+export const almostEqual = (a, b) => Math.abs(a - b) < 0.000001;
 
 // Bilinear interpolation between two data arrays
-function interpolateArray(data1, data2, w) {
+export function interpolateArray(data1, data2, w) {
     if (almostEqual(w, 0)) {
-        return _.clone(data1);
+        return clone(data1);
     }
     if (almostEqual(w, 1)) {
-        return _.clone(data2);
+        return clone(data2);
     }
 
     const data = [];
     for (let lev = 0; lev < data1.length; ++lev) {
         // Some levels may be missing (corrupted data) except ground level
-        let d2 = lev == 0 ? data2[0] : data2.find(d => d.level == data1[lev].level);
+        const d2 = lev == 0 ? data2[0] : data2.find(d => d.level == data1[lev].level);
         if (!d2) {
             continue;
         }
@@ -38,12 +38,12 @@ function interpolateArray(data1, data2, w) {
 }
 
 // Bilinear interpolation between two data points
-function interpolatePoint(point1, point2, w) {
+export function interpolatePoint(point1, point2, w) {
     if (almostEqual(w, 0)) {
-        return _.clone(point1);
+        return clone(point1);
     }
     if (almostEqual(w, 1)) {
-        return _.clone(point2);
+        return clone(point2);
     }
 
     const interp = {};
@@ -56,7 +56,7 @@ function interpolatePoint(point1, point2, w) {
 }
 
 // Returns intersection of two line segments
-function intersection(line1, line2) {
+export function intersection(line1, line2) {
     const [[x1, y1], [x2, y2]] = line1;
     const [[x3, y3], [x4, y4]] = line2;
 
@@ -64,20 +64,20 @@ function intersection(line1, line2) {
         return null; // One of the lines has zero length
     }
 
-    var d = (y2 - y1) * (x4 - x3) - (x2 - x1) * (y4 - y3);
+    const d = (y2 - y1) * (x4 - x3) - (x2 - x1) * (y4 - y3);
     if (!d) {
         return null;
     } // Lines are parallel
 
-    var t = ((x2 - x1) * (y3 - y1) + (y2 - y1) * (x1 - x3)) / d;
+    const t = ((x2 - x1) * (y3 - y1) + (y2 - y1) * (x1 - x3)) / d;
     if (t < 0 || t > 1) {
         return null;
     } // Intersection is out of line1
 
-    var px = x3 + t * (x4 - x3);
-    var py = y3 + t * (y4 - y3);
+    const px = x3 + t * (x4 - x3);
+    const py = y3 + t * (y4 - y3);
 
-    var s = x2 - x1 ? (px - x1) / (x2 - x1) : (py - y1) / (y2 - y1);
+    const s = x2 - x1 ? (px - x1) / (x2 - x1) : (py - y1) / (y2 - y1);
     if (s < 0 || s > 1) {
         return null;
     } // Intersetion is out of line2
@@ -86,7 +86,7 @@ function intersection(line1, line2) {
 }
 
 // Returns first intersection of polyline straight line
-function dataIntersection(line, polyline, getPoint) {
+export function dataIntersection(line, polyline, getPoint) {
     for (let i = 0; i < polyline.length - 1; ++i) {
         const pt = intersection(line, [getPoint(polyline[i]), getPoint(polyline[i + 1])]);
         if (pt) {
@@ -98,7 +98,7 @@ function dataIntersection(line, polyline, getPoint) {
 }
 
 // Inspired by airgram.mjs:windMark()
-function addWindBarb(g, x, y, dir, speed) {
+export function addWindBarb(g, x, y, dir, speed) {
     const scale = 20;
     const windBarb = g
         .append('g')
@@ -118,7 +118,7 @@ function addWindBarb(g, x, y, dir, speed) {
             .attr('y2', -scale);
 
         const d = 0.15 * scale;
-        var dy = -scale;
+        let dy = -scale;
 
         // Shift 5 kn barb
         if (knots5 == 1) {
@@ -170,13 +170,3 @@ function addWindBarb(g, x, y, dir, speed) {
             .attr('r', 0.05 * scale);
     }
 }
-
-export default {
-    validateData,
-    almostEqual,
-    interpolateArray,
-    interpolatePoint,
-    intersection,
-    dataIntersection,
-    addWindBarb,
-};
